+++ build.gradle.kts
@@ -11,6 +11,8 @@ repositories {
 
 dependencies {
     testImplementation(kotlin("test"))
+    testImplementation("junit:junit:4.13.1")
+
 }
 
 tasks.test {
