+++ docs/daily/2024-12-19.md
@@ -1,0 +1,16 @@
+# 2024-12-19
+
+## 📚 오늘 배운 내용
+- 힙 영역과 스택 영역
+- 생성자의 매개변수 순서에 관한 내용
+- 생성자 매개변수 초기값 설정
+- companion object
+
+## 💻 예제 코드
+<!-- 실습한 코드나 예제를 추가 -->
+
+## 🔍 참고 자료
+- 
+
+## ❓ 궁금한 점
+-
