+++ docs/topics/tempTopics/2024-12-19_topic.md
@@ -1,0 +1,19 @@
+## 힙영역과 스택영역
+### 힙 영역
+- 메모리의 할당과 소멸이 동적으로 이루어지며 사용자의 코드에 따라 유동적이다.
+- 정확히는 할당만으로 소멸은 GC가 대신 해준다.
+### 스택영역
+- 일반 지역 변수들이 스택 영역 메모리 공간에 올라온다.
+- 스택의 구조에 따라서 메모리의 할당과 소멸이 이루어진다.
+
+## 생성자의 매개 변수 순서
+- 다른 언어에서는 생성자의 매개 변수 순서를 지켜서 인스턴스를 생성해야하지만  
+코틀린에서는 그 부분을 신경쓰지 않아도 된다.
+
+## 생성자 매개변수의 초기값 설정
+- 생성자 매개변수를 정의해줄 때 초기값을 설정해서  
+인스턴스 생성 시에 아무런 값도 입력하지 않을 수 있다.
+
+## companion object
+- 인스턴스 영역이 아니라 클래스 영역이라고 보면 된다.
+- 다른 객체지향 문법의 클래스 내 필드의 static 선언과 동일한 특성을 갖는다.
